Track number of open connections and lost packets to not quit while awaiting responses
The problem was valid transformations could be lost if the program terminates
while awaiting replies to previously sent packets. For example, if two
individual packets both with valid transformations were sent consecutively,
the program sends the first packet and awaits the first reply. Then, after
the waiting timeout it will send the second request. But if the second request
was sent sooner than the first reply was received, the program terminates since
the `todo` list is empty, failing to await the second reply packet.